+++ src/crawlers/cheerio_crawler.js
@@ -7,8 +7,8 @@ import { WritableStream } from 'htmlparser2/lib/WritableStream';
 import * as iconv from 'iconv-lite';
 import ow from 'ow';
 import * as util from 'util';
+import { TimeoutError } from 'got-scraping';
 import { BASIC_CRAWLER_TIMEOUT_BUFFER_SECS } from '../constants';
-import { TimeoutError } from '../errors';
 import { addTimeoutToPromise, parseContentTypeFromResponse } from '../utils';
 import * as utilsRequest from '../utils_request'; // eslint-disable-line import/no-duplicates
 import BasicCrawler from './basic_crawler'; // eslint-disable-line import/no-duplicates
@@ -554,10 +554,10 @@ class CheerioCrawler extends BasicCrawler {
         }
 
         const opts = this._getRequestOptions(request, session, proxyUrl);
-        let responseStream;
+        let responseWithStream;
 
         try {
-            responseStream = await utilsRequest.requestAsBrowser(opts);
+            responseWithStream = await utilsRequest.requestAsBrowser(opts);
         } catch (e) {
             if (e instanceof TimeoutError) {
                 this._handleRequestTimeout(session);
@@ -566,7 +566,7 @@ class CheerioCrawler extends BasicCrawler {
             }
         }
 
-        return responseStream;
+        return responseWithStream;
     }
 
     /**
@@ -705,7 +705,9 @@ class CheerioCrawler extends BasicCrawler {
             });
             const parser = new WritableStream(domHandler, { decodeEntities: true });
             parser.on('error', reject);
-            response.on('error', reject).pipe(parser);
+            response
+                .on('error', reject)
+                .pipe(parser);
         });
     }